+++ src/core/functions/commentTrigger.ts
@@ -1,8 +1,8 @@
 import { A_ANY } from "@/@types/ast";
 import { argumentParser, currentTime } from "@/context";
 import { addHandler } from "@/commentHandler";
-import { IrFunction } from "@/core/functions/index";
 import { execute } from "@/core/coreContext";
+import { IrFunction } from "@/@types/core/functions";
 
 const processCommentTrigger: IrFunction = (script, scopes) => {
   const args = argumentParser(
