+++ src/core/functions/drawShape.ts
@@ -1,6 +1,7 @@
 import { A_CallExpression, T_scope } from "@/@types/ast";
 import { argumentParser, context } from "@/context";
 import { IrShape } from "@/objects/shape";
+import { IrFunction } from "@/@types/core/functions";
 
 /**
  * @関数
@@ -8,7 +9,10 @@ import { IrShape } from "@/objects/shape";
  * @param script
  * @param scopes
  */
-const processDrawShape = (script: A_CallExpression, scopes: T_scope[]) => {
+const processDrawShape: IrFunction = (
+  script: A_CallExpression,
+  scopes: T_scope[]
+) => {
   const args = argumentParser(script.arguments, scopes, [
     "x",
     "y",
