+++ src/core/functions/dump.ts
@@ -1,5 +1,6 @@
 import { A_CallExpression, T_scope } from "@/@types/ast";
 import { execute } from "@/core/coreContext";
+import { IrFunction } from "@/@types/core/functions";
 
 /**
  * @非標準
@@ -8,7 +9,10 @@ import { execute } from "@/core/coreContext";
  * @param script
  * @param scopes
  */
-const processDump = (script: A_CallExpression, scopes: T_scope[]) => {
+const processDump: IrFunction = (
+  script: A_CallExpression,
+  scopes: T_scope[]
+) => {
   for (const argument of script.arguments) {
     console.debug("%cdump", "background:green;", execute(argument, scopes));
   }
