+++ src/core/functions/playStartTime.ts
@@ -1,5 +1,8 @@
 const time = new Date().getTime();
+import { IrFunction } from "@/@types/core/functions";
 
-const processPlayStartTime = () => {
+const processPlayStartTime: IrFunction = () => {
   return time;
 };
+
+export { processPlayStartTime };
