+++ src/core/functions/screen.ts
@@ -1,11 +1,12 @@
 import { config } from "@/definition/config";
 import { isWide } from "@/context";
+import { IrFunction } from "@/@types/core/functions";
 
-const processScreenWidth = () => {
+const processScreenWidth: IrFunction = () => {
   return config.stageWidth[isWide ? "full" : "default"];
 };
 
-const processScreenHeight = () => {
+const processScreenHeight: IrFunction = () => {
   return config.stageHeight;
 };
