+++ src/core/functions/timer.ts
@@ -2,8 +2,12 @@ import { A_ANY, A_CallExpression, T_scope } from "@/@types/ast";
 import { argumentParser } from "@/context";
 import { addQueue } from "@/queue";
 import { execute } from "@/core/coreContext";
+import { IrFunction } from "@/@types/core/functions";
 
-const processTimer = (script: A_CallExpression, scopes: T_scope[]) => {
+const processTimer: IrFunction = (
+  script: A_CallExpression,
+  scopes: T_scope[]
+) => {
   const args = argumentParser(
     script.arguments,
     scopes,
