+++ src/core/prototype/Object/index.ts
@@ -1,25 +1,10 @@
-import { A_CallExpression, A_MemberExpression, T_scope } from "@/@types/ast";
 import { processDef } from "@/core/prototype/Object/def";
 import { processGetSlot } from "@/core/prototype/Object/getSlot";
 import { processSetSlot } from "@/core/prototype/Object/setSlot";
 import { processClone } from "@/core/prototype/Object/clone";
+import { PrototypeFunctions } from "@/@types/core/prototype";
 
-export type PrototypeFunction = (
-  script: A_CallExpression & { callee: A_MemberExpression },
-  scopes: T_scope[],
-  object: { [key: string]: unknown }
-) => unknown;
-
-export type PrototypeFunctions = {
-  [key: string]:
-    | {
-        func: PrototypeFunction;
-        condition: (object: unknown, script: A_CallExpression) => boolean;
-      }
-    | PrototypeFunction;
-};
-
-const prototypeFunctions = {
+const prototypeFunctions: PrototypeFunctions = {
   def: processDef,
   getSlot: processGetSlot,
   setSlot: processSetSlot,
